Add the source for the SKL version of original OP2 compiler from ETHZ.
The compiler sources in this change are essentially the original
compiler, but the modularity was changed many years ago to accommodate
a 32-bit x86 code generator which generated more code that would fit
in the original buffers.  A few modifications remain from the 32-bit
x86 project, but that doesn't detract from the compilers overall
usefulness.